Rename misnamed file
